version 1.0
I added an array list because then i was able to get the hands and from
the hands the index finger and the index finger to the tip in order for
something to be drawn on the screen. Then I made it so that whenever
the index finger moves a circle is drawn and I used a map function so
that it is easier for the finger to draw the circles. (i had to make a
new repository because the other one was not letting me insert my final
code and it was not letting me comit.)